use helper to centralize business logic
