+++ src/main/kotlin/it/pagopa/ecommerce/eventdispatcher/queues/TransactionAuthorizationRequestedQueueConsumer.kt
@@ -4,6 +4,7 @@ import com.azure.core.util.BinaryData
 import com.azure.core.util.serializer.TypeReference
 import com.azure.spring.messaging.AzureHeaders
 import com.azure.spring.messaging.checkpoint.Checkpointer
+import io.vavr.control.Either
 import it.pagopa.ecommerce.commons.documents.v2.TransactionAuthorizationRequestedEvent as TransactionAuthorizationRequestedEventV2
 import it.pagopa.ecommerce.commons.queues.QueueEvent
 import it.pagopa.ecommerce.commons.queues.StrictJsonSerializerProvider
@@ -56,11 +57,23 @@ class TransactionAuthorizationRequestedQueueConsumer(
     @Header(AzureHeaders.CHECKPOINTER) checkPointer: Checkpointer
   ): Mono<Unit> {
     val eventWithTracingInfo = parseEvent(payload)
+
     return eventWithTracingInfo
-      .flatMap {
-        logger.debug(
-          "Event {} with tracing info {} dispatched to V2 handler", it.event, it.tracingInfo)
-        queueConsumerV2.messageReceiver(it, checkPointer)
+      .flatMap { e ->
+        when (e.event) {
+          is TransactionAuthorizationRequestedEventV2 -> {
+            logger.debug(
+              "Event {} with tracing info {} dispatched to V2 handler", e.event, e.tracingInfo)
+            queueConsumerV2.messageReceiver(Either.left(e), checkPointer)
+          }
+          else -> {
+            logger.error(
+              "Event {} with tracing info {} cannot be dispatched to any know handler",
+              e.event,
+              e.tracingInfo)
+            Mono.error(InvalidEventException(payload, null))
+          }
+        }
       }
       .onErrorResume(InvalidEventException::class.java) {
         logger.error("Invalid input event", it)
